Dissolve preSteps as they are Conditions
Both the ReadInstance and HandleDelete steps are really conditions that
branching the reconciliation to two totally different path. So we should
not try to handle them as steps with a special effect of skipping some
other steps.

At the moment read instance and handle delete are inlined for
simplicity. IF we need to handle a lot of conditional paths later then
we can reconsider building some top level branching concept.

The two path defined by the handle delete condition will eventually need
two set of steps. The currently existing "normal" steps and a different
"cleanup" steps. As the latter is empty today we will handle that when
adding a later example.