Implement proper instance save
This dissolved the postStep internal concept as so far only save
instance is found as a post step. We can introduce it back later if
needed but for simplicity we inlined the instance saving to handleReq.

To properly save the instance changes we need two patch calls one for
the instance metadata (later used for self finalizer) and one for the
status fields. Also we use patch instead of update as updating the
instance metadata would try to update everything that is not status
which would lead to frequent conflicts.

To be able to calculate a patch diff we need to make a copy of the
instance before we let the reconcile steps modifying it.