Move Ready Condition recalculation to a step
We only need to recalculate ready condition if reconciled instance has
conditions. Also there can be instances that does not follow the pattern
to define the Ready condition state based on all of the other
conditions. By moving the Ready recalculation to a step it can be added
optionally to the reconciler. However this recalculation always needs
to be done regardless of the normal step results, i.e. if a step fail we
should not skip the condition recalculation step. So the post step
concept is re-introduced. Post steps are always executed after the
normal steps' execution finished (successfully or unsuccessfully).